+++ metadata.rb
@@ -6,7 +6,7 @@ description 'Installs/Configures node.js'
 source_url 'https://github.com/redguide/nodejs'
 issues_url 'https://github.com/redguide/nodejs/issues'
 chef_version '>= 14'
-version '7.2.0'
+version '7.3.0'
 
 depends 'ark', '>= 2.0.2'
 depends 'chocolatey', '>= 3.0'